+++ src/components/Button/Button.module.scss
@@ -7,7 +7,7 @@
 
 
 
-.basicButtonWrapper{
+.buttonWrapper{
   background-color: #E8AE1E;
   border-radius: 30px;
   border-color: transparent;
